HW7: is completed (#8)
* HW7: is completed

* HW7: test

* HW7: test

* HW7: test

* HW7: test

* HW7: fix

* HW7: fix

* HW7: redundant test removed

* HW7: new test added | tests refactored

* HW7: fix

* HW7: self review

* HW7: special files processing added

* HW7: spec files check changed | tests extended

* HW7: test

* HW7: fix

* HW7: fix

* HW7: debug

* HW7: symlinks processing support

+++ hw07_file_copying/test.sh
@@ -1,25 +1,48 @@
 #!/usr/bin/env bash
 set -xeuo pipefail
 
+TO=out.txt
+FROM=testdata/input.txt
+
 go build -o go-cp
 
-./go-cp -from testdata/input.txt -to out.txt
-cmp out.txt testdata/out_offset0_limit0.txt
+./go-cp -from $FROM -to $TO
+cmp $TO testdata/out_offset0_limit0.txt
 
-./go-cp -from testdata/input.txt -to out.txt -limit 10
-cmp out.txt testdata/out_offset0_limit10.txt
+./go-cp -from $FROM -to $TO -limit 10
+cmp $TO testdata/out_offset0_limit10.txt
 
-./go-cp -from testdata/input.txt -to out.txt -limit 1000
-cmp out.txt testdata/out_offset0_limit1000.txt
+./go-cp -from $FROM -to $TO -limit 1000
+cmp $TO testdata/out_offset0_limit1000.txt
 
-./go-cp -from testdata/input.txt -to out.txt -limit 10000
-cmp out.txt testdata/out_offset0_limit10000.txt
+./go-cp -from $FROM -to $TO -limit 10000
+cmp $TO testdata/out_offset0_limit10000.txt
 
-./go-cp -from testdata/input.txt -to out.txt -offset 100 -limit 1000
-cmp out.txt testdata/out_offset100_limit1000.txt
+./go-cp -from $FROM -to $TO -offset 100 -limit 1000
+cmp $TO testdata/out_offset100_limit1000.txt
 
-./go-cp -from testdata/input.txt -to out.txt -offset 6000 -limit 1000
-cmp out.txt testdata/out_offset6000_limit1000.txt
+./go-cp -from $FROM -to $TO -offset 6000 -limit 1000
+cmp $TO testdata/out_offset6000_limit1000.txt
+
+./go-cp -from testdata/empty.txt -to $TO
+cmp $TO testdata/out_empty.txt
+
+rm -f $TO
+./go-cp -from $FROM -to $TO -offset 100000
+if [ -f $TO ]; then
+  exit 1
+fi
+./go-cp -from /dev/urandom -to $TO
+if [ -f $TO ]; then
+  exit 1
+fi
+
+ln -sf $FROM symlink
+./go-cp -from symlink -to $TO
+if [ -f $TO ]; then
+  exit 1
+fi
+rm -f symlink
 
 rm -f go-cp out.txt
 echo "PASS"
